Fix: read zv from string with scientific notation

+++ source/module_cell/read_pp_upf201.cpp
@@ -292,7 +292,7 @@ void Pseudopot_upf::read_pseudo_upf201_header(std::ifstream& ifs)
         }
         else if (name[ip] == "z_valence")
         {
-            zp = atoi(val[ip].c_str());
+            zp = atof(val[ip].c_str());
         }
         else if (name[ip] == "total_psenergy")
         {
